requirements.txt
streamlit
google-genai
google-cloud-aiplatform
python-docx
graphviz
pydantic